gpu: pvr: replace boolean by flags in blits complete query
In preparation for adding third mode of query. In DRM one
has a specific IOCTL for requesting the DRM kernel driver to
issue events when a particular vertical blanking period
passes. Here one aims to introduce similar IOCTL for requesting
PVR kernel driver to issue events whenever a particular frame
gets rendered in full. And instead of introducing entirely new
IOCTL, one instead aims to extend an existing call with a new
mode. This is accomplished by replacing the boolean selector
with an integer.

Signed-off-by: Topi Pohjolainen <topi.pohjolainen@nokia.com>
Signed-off-by: Imre Deak <imre.deak@nokia.com>